refactor: split review list rendering
